--- src/main/java/com/example/springbootbackend/SpringBootBackendApplication.java
@@ -1,13 +1,0 @@
-package com.example.springbootbackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringBootBackendApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBootBackendApplication.class, args);
-    }
-
-}
