--- src/main/java/com/example/springbootbackend/models/user/ERole.java
@@ -1,7 +1,0 @@
-package com.example.springbootbackend.models.user;
-
-public enum ERole {
-    ROLE_USER,
-    ROLE_MODERATOR,
-    ROLE_ADMIN
-}
