--- src/main/java/com/example/springbootbackend/payload/request/LoginRequest.java
@@ -1,27 +1,0 @@
-package com.example.springbootbackend.payload.request;
-
-import javax.validation.constraints.NotBlank;
-
-public class LoginRequest {
-    @NotBlank
-    private String username;
-
-    @NotBlank
-    private String password;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
