--- src/main/java/com/example/springbootbackend/payload/response/MessageResponse.java
@@ -1,18 +1,0 @@
-package com.example.springbootbackend.payload.response;
-
-public class MessageResponse {
-    private String message;
-
-    public MessageResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
-
