--- src/main/java/com/example/springbootbackend/repository/Repository.java
@@ -1,13 +1,0 @@
-package com.example.springbootbackend.repository;
-
-import java.util.List;
-
-public interface Repository<T> {
-    List<T> findAll();
-
-    T findById(Long id);
-
-    void save(T model);
-
-    void remove(Long id);
-}
