--- src/main/java/com/example/springbootbackend/repository/RoleRepository.java
@@ -1,14 +1,0 @@
-package com.example.springbootbackend.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.springbootbackend.models.user.ERole;
-import com.example.springbootbackend.models.user.Role;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName(ERole name);
-}
