--- src/main/java/com/example/springbootbackend/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.springbootbackend.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import com.example.springbootbackend.models.user.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-
-    Boolean existsByUsername(String username);
-
-    Boolean existsByEmail(String email);
-}
