--- src/main/java/com/example/springbootbackend/repository/post/PostRepository.java
@@ -1,7 +1,0 @@
-package com.example.springbootbackend.repository.post;
-
-import com.example.springbootbackend.models.Post;
-import com.example.springbootbackend.repository.Repository;
-
-public interface PostRepository extends Repository<Post> {
-}
