--- src/main/java/com/example/springbootbackend/security/services/post/PostService.java
@@ -1,15 +1,0 @@
-package com.example.springbootbackend.security.services.post;
-
-import com.example.springbootbackend.models.Post;
-
-import java.util.List;
-
-public interface PostService {
-    List<Post> findAll();
-
-    Post findById(Long id);
-
-    void save (Post post);
-
-    void remove(Long id);
-}
