+++ src/main/java/com/horizonevent/project/SpringBootBackendApplication.java
@@ -1,0 +1,13 @@
+package com.horizonevent.project;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootBackendApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootBackendApplication.class, args);
+    }
+
+}
