+++ src/main/java/com/horizonevent/project/models/user/ERole.java
@@ -1,0 +1,7 @@
+package com.horizonevent.project.models.user;
+
+public enum ERole {
+    ROLE_USER,
+    ROLE_MODERATOR,
+    ROLE_ADMIN
+}
