+++ src/main/java/com/horizonevent/project/payload/request/LoginRequest.java
@@ -1,0 +1,27 @@
+package com.horizonevent.project.payload.request;
+
+import javax.validation.constraints.NotBlank;
+
+public class LoginRequest {
+    @NotBlank
+    private String username;
+
+    @NotBlank
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
