+++ src/main/java/com/horizonevent/project/payload/response/MessageResponse.java
@@ -1,0 +1,18 @@
+package com.horizonevent.project.payload.response;
+
+public class MessageResponse {
+    private String message;
+
+    public MessageResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
+
