+++ src/main/java/com/horizonevent/project/repository/Repository.java
@@ -1,0 +1,13 @@
+package com.horizonevent.project.repository;
+
+import java.util.List;
+
+public interface Repository<T> {
+    List<T> findAll();
+
+    T findById(Long id);
+
+    void save(T model);
+
+    void remove(Long id);
+}
