+++ src/main/java/com/horizonevent/project/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package com.horizonevent.project.repository;
+
+import java.util.Optional;
+
+import com.horizonevent.project.models.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+
+    Boolean existsByUsername(String username);
+
+    Boolean existsByEmail(String email);
+}
