+++ src/main/java/com/horizonevent/project/repository/post/PostRepository.java
@@ -1,0 +1,7 @@
+package com.horizonevent.project.repository.post;
+
+import com.horizonevent.project.models.Post;
+import com.horizonevent.project.repository.Repository;
+
+public interface PostRepository extends Repository<Post> {
+}
