+++ src/main/java/com/horizonevent/project/security/services/post/PostServiceImpl.java
@@ -1,0 +1,34 @@
+package com.horizonevent.project.security.services.post;
+
+import com.horizonevent.project.models.Post;
+import com.horizonevent.project.repository.post.PostRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class PostServiceImpl implements PostService {
+    @Autowired
+    private PostRepository postRepository;
+
+    @Override
+    public List<Post> findAll() {
+        return postRepository.findAll();
+    }
+
+    @Override
+    public Post findById(Long id) {
+        return postRepository.findById(id);
+    }
+
+    @Override
+    public void save(Post post) {
+        postRepository.save(post);
+    }
+
+    @Override
+    public void remove(Long id) {
+        postRepository.remove(id);
+    }
+}
