+++ src/main/resources/application.properties
@@ -16,6 +16,5 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 # App Properties
-bezkoder.app.jwtSecret= horizonEventSecretKey
+bezkoder.app.jwtSecret= bezkoder
 bezkoder.app.jwtExpirationMs= 86400000
-fileLink=D:\\Codegym\\Season 4\\Project\\picture
