--- src/test/java/com/example/springbootbackend/SpringBootBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.springbootbackend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringBootBackendApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
