+++ src/test/java/com/horizonevent/project/SpringBootBackendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.horizonevent.project;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBootBackendApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
